requirements.txt
numpy==1.21.1
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
visdom==0.1.8.9
Pillow==8.3.1